Add files via upload
Chances are if you can make an engine run, you can figure out how to splice this into a gasoline driven engine with 'points & coil' ignition.  The points are now a sensor, telling this when to provide a signal to turn on & off the coil via a transistor.  Or use a coil on plug style ignitor for all i care.
A useful upgrade to older engines you're 'souping up' or just re-fitting a sensor to worn out distributors with damaged or missing parts to restore full function (or add new function) again.
Chances are an electronic fuel injection system can be crafted to plug right into this with a few more lines of code.  And a camshaft sensor.